+++ simulator/classes/__init__.py
@@ -4,3 +4,4 @@ from .finale import Finale
 from .nom_ceremony import NominationCeremony
 from .veto_players import VetoPlayers
 from .veto_ceremony import VetoCeremony
+from .week import Week
