+++ simulator/factories.py
@@ -2,7 +2,6 @@ from simulator.models import (
     Contestant,
     Houseguest,
     Game,
-    Week,
 )
 from faker import Faker
 
@@ -30,28 +29,3 @@ class HouseguestFactory(factory.django.DjangoModelFactory):
     name = fake.first_name()
     game = factory.SubFactory(GameFactory)
 
-
-class WeekFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = Week
-
-    hoh = factory.SubFactory(HouseguestFactory)
-    evicted = factory.SubFactory(HouseguestFactory)
-    pov = factory.SubFactory(HouseguestFactory)
-    number = 1
-
-    @factory.post_generation
-    def initial_nominees(self, create, extracted, **kwargs):
-        if not create:
-            return
-        if extracted:
-            for nominee in extracted:
-                self.initial_nominees.add(nominee)
-
-    @factory.post_generation
-    def final_nominees(self, create, extracted, **kwargs):
-        if not create:
-            return
-        if extracted:
-            for nominee in extracted:
-                self.final_nominees.add(nominee)
