+++ simulator/models/__init__.py
@@ -1,4 +1,3 @@
-from .week import Week
 from .game import Game
 from .relationship import Relationship
 from .houseguest import Houseguest
