+++ simulator/unit_tests/test_game.py
@@ -1,11 +1,11 @@
 import pytest
 
-from ..classes import Competition, NominationCeremony, VetoCeremony, EvictionCeremony, Finale
+from ..classes import Competition, NominationCeremony, VetoCeremony, EvictionCeremony, Finale, Week
 from ..models import (
     Game,
     Houseguest,
 )
-from ..factories import HouseguestFactory, GameFactory, WeekFactory
+from ..factories import HouseguestFactory, GameFactory
 
 
 class TestGame:
@@ -306,29 +306,32 @@ class TestGame:
         # Wk 3: HOH - 0, Nom - 2 and 1, POV - 2, Final 1 and 5, Evicted: 5
         # Final 3: 0, 1, 2
 
-        wk1 = WeekFactory(
+        wk1 = Week(
             number=1,
             hoh=hgs[0],
             initial_nominees=[hgs[2], hgs[3]],
             pov=hgs[4],
             final_nominees=[hgs[2], hgs[3]],
             evicted=hgs[3],
+            vote_count=[2,1]
         )
-        wk2 = WeekFactory(
+        wk2 = Week(
             number=2,
             hoh=hgs[2],
             initial_nominees=[hgs[0], hgs[4]],
             pov=hgs[0],
             final_nominees=[hgs[4], hgs[1]],
             evicted=hgs[4],
+            vote_count=[2,0]
         )
-        wk3 = WeekFactory(
+        wk3 = Week(
             number=3,
             hoh=hgs[0],
             initial_nominees=[hgs[2], hgs[1]],
             pov=hgs[2],
             final_nominees=[hgs[1], hgs[5]],
             evicted=hgs[5],
+            vote_count=[1, 0]
         )
 
         wks = [wk1, wk2, wk3]
